transition to cmake build + reorganization of files
