--- debug/log.txt
@@ -1,143 +1,0 @@
-findBracket init:    10.1 -> 6857.5    10.35 -> 7642.69    10.35 -> 7642.69
-findBracket (C):    9.85 -> 6134.41    10.1 -> 6857.5    10.35 -> 7642.69
-findBracket (D):    9.35 -> 4861.23    9.85 -> 6134.41    10.1 -> 6857.5
-findBracket (D):    8.35 -> 2918.43    9.35 -> 4861.23    9.85 -> 6134.41
-findBracket (D):    6.35 -> 819.248    8.35 -> 2918.43    9.35 -> 4861.23
-findBracket (D):    2.35 -> 3.32151    6.35 -> 819.248    8.35 -> 2918.43
-findBracket (D):    -5.65 -> 1955.63    2.35 -> 3.32151    6.35 -> 819.248
-findBracket:    -5.65 -> 1955.63    2.35 -> 3.32151    6.35 -> 819.248
-
-
-=========================================================================
-
-
-parabgoldMinimization:    -5.65 -> 1955.63    2.35 -> 3.32151    6.35 -> 819.248
-parabgoldMinimization:    0.855357 -> 0.000437719    1.00296 -> 7.68088e-11    1.12047 -> 0.000210607
-
-
-=========================================================================
-
-
-
-Begin ConjGrad::findMin() procedure
-
-Current position and target value:
-x0 = -2    x1 = 1    x2 = 0
-f = 0 +- 0
-
-Raw gradient:
-g0 = -108 +- 1e-06    g1 = 62.5 +- 1e-06    g2 = -0.5 +- 1e-06
-
-Conjugated Vectors:
-g0 = 108    g1 = -62.5    g2 = 0.5
-
-findBracket init:    0 -> 120.125    0.25 -> 361451    0.25 -> 361451
-findBracket (C):    -0.25 -> 917923    0 -> 120.125    0.25 -> 361451
-findBracket:    -0.25 -> 917923    0 -> 120.125    0.25 -> 361451
-parabgoldMinimization:    -0.25 -> 917923    0 -> 120.125    0.25 -> 361451
-parabgoldMinimization:    0.0316977 -> 0.159563    0.0317583 -> 0.159488    0.0318446 -> 0.159626
-Current position and target value:
-x0 = 1.4299    x1 = -0.984895    x2 = 0.0158792
-f = 0.159488 +- 1e-05
-
-
-
-ConjGrad::findMin() Step 1
-
-Raw gradient:
-g0 = 0.317801 +- 1e-06    g1 = 0.546699 +- 1e-06    g2 = -0.453859 +- 1e-06
-
-Conjugated vectors:
-g0 = -0.313599    g1 = -0.549131    g2 = 0.453879
-
-findBracket init:    0 -> 0.159488    0.25 -> 0.0545163    0.25 -> 0.0545163
-findBracket (A):    0 -> 0.159488    0.25 -> 0.0545163    0.5 -> 0.0132851
-findBracket (B):    0.25 -> 0.0545163    0.5 -> 0.0132851    1 -> 0.000185114
-findBracket (B):    0.5 -> 0.0132851    1 -> 0.000185114    2 -> 0.149373
-findBracket:    0.5 -> 0.0132851    1 -> 0.000185114    2 -> 0.149373
-parabgoldMinimization:    0.5 -> 0.0132851    1 -> 0.000185114    2 -> 0.149373
-parabgoldMinimization:    1 -> 0.000185114    1.08432 -> 0.000106832    1.19098 -> 0.000392515
-Current position and target value:
-x0 = 1.08986    x1 = -1.58033    x2 = 0.508027
-f = 0.000106832 +- 1e-05
-
-
-
-ConjGrad::findMin() Step 2
-
-Raw gradient:
-g0 = 0.00290217 +- 1e-06    g1 = -0.00207315 +- 1e-06    g2 = 2.06899e-06 +- 1e-06
-
-Conjugated vectors:
-g0 = -0.00290217    g1 = 0.00207315    g2 = -2.06862e-06
-
-findBracket init:    0 -> 0.000106832    0.25 -> 0.000103687    0.25 -> 0.000103687
-findBracket init:    0 -> 0.000106832    0.375 -> 0.000102142    0.375 -> 0.000102142
-findBracket init:    0 -> 0.000106832    0.5625 -> 9.98561e-05    0.5625 -> 9.98561e-05
-findBracket init:    0 -> 0.000106832    0.84375 -> 9.65007e-05    0.84375 -> 9.65007e-05
-findBracket init:    0 -> 0.000106832    1.26562 -> 9.1629e-05    1.26562 -> 9.1629e-05
-findBracket init:    0 -> 0.000106832    1.89844 -> 8.46742e-05    1.89844 -> 8.46742e-05
-findBracket init:    0 -> 0.000106832    2.84766 -> 7.50029e-05    2.84766 -> 7.50029e-05
-findBracket init:    0 -> 0.000106832    4.27148 -> 6.20988e-05    4.27148 -> 6.20988e-05
-findBracket (A):    0 -> 0.000106832    4.27148 -> 6.20988e-05    8.54297 -> 3.32972e-05
-findBracket (B):    0 -> 0.000106832    4.27148 -> 6.20988e-05    17.0859 -> 6.70011e-06
-findBracket (B):    4.27148 -> 6.20988e-05    17.0859 -> 6.70011e-06    42.7148 -> 1.362e-06
-findBracket (B):    4.27148 -> 6.20988e-05    17.0859 -> 6.70011e-06    93.9727 -> 0.00129011
-findBracket:    4.27148 -> 6.20988e-05    17.0859 -> 6.70011e-06    93.9727 -> 0.00129011
-parabgoldMinimization:    4.27148 -> 6.20988e-05    17.0859 -> 6.70011e-06    93.9727 -> 0.00129011
-parabgoldMinimization:    4.27148 -> 6.20988e-05    34.0509 -> 1.9441e-08    93.9727 -> 0.00129011
-Current position and target value:
-x0 = 0.991036    x1 = -1.50973    x2 = 0.507957
-f = 1.9441e-08 +- 1e-05
-
-
-
-ConjGrad::findMin() Step 3
-
-Raw gradient:
-g0 = -2.88119e-06 +- 1e-06    g1 = -3.68869e-06 +- 1e-06    g2 = 2.015e-06 +- 1e-06
-
-Conjugated vectors:
-g0 = 2.88119e-06    g1 = 3.68869e-06    g2 = -2.015e-06
-
-findBracket init:    0 -> 1.9441e-08    0.25 -> 1.94345e-08    0.25 -> 1.94345e-08
-findBracket init:    0 -> 1.9441e-08    0.375 -> 1.94313e-08    0.375 -> 1.94313e-08
-findBracket init:    0 -> 1.9441e-08    0.5625 -> 1.94264e-08    0.5625 -> 1.94264e-08
-findBracket init:    0 -> 1.9441e-08    0.84375 -> 1.94191e-08    0.84375 -> 1.94191e-08
-findBracket init:    0 -> 1.9441e-08    1.26562 -> 1.94082e-08    1.26562 -> 1.94082e-08
-findBracket init:    0 -> 1.9441e-08    1.89844 -> 1.93917e-08    1.89844 -> 1.93917e-08
-findBracket init:    0 -> 1.9441e-08    2.84766 -> 1.93672e-08    2.84766 -> 1.93672e-08
-findBracket init:    0 -> 1.9441e-08    4.27148 -> 1.93303e-08    4.27148 -> 1.93303e-08
-findBracket init:    0 -> 1.9441e-08    6.40723 -> 1.92752e-08    6.40723 -> 1.92752e-08
-findBracket init:    0 -> 1.9441e-08    9.61084 -> 1.91927e-08    9.61084 -> 1.91927e-08
-findBracket init:    0 -> 1.9441e-08    14.4163 -> 1.90694e-08    14.4163 -> 1.90694e-08
-findBracket init:    0 -> 1.9441e-08    21.6244 -> 1.88856e-08    21.6244 -> 1.88856e-08
-findBracket init:    0 -> 1.9441e-08    32.4366 -> 1.86126e-08    32.4366 -> 1.86126e-08
-findBracket init:    0 -> 1.9441e-08    48.6549 -> 1.82086e-08    48.6549 -> 1.82086e-08
-findBracket init:    0 -> 1.9441e-08    72.9823 -> 1.76153e-08    72.9823 -> 1.76153e-08
-findBracket init:    0 -> 1.9441e-08    109.473 -> 1.67534e-08    109.473 -> 1.67534e-08
-findBracket init:    0 -> 1.9441e-08    164.21 -> 1.55215e-08    164.21 -> 1.55215e-08
-findBracket init:    0 -> 1.9441e-08    246.315 -> 1.38055e-08    246.315 -> 1.38055e-08
-findBracket init:    0 -> 1.9441e-08    369.473 -> 1.15094e-08    369.473 -> 1.15094e-08
-findBracket init:    0 -> 1.9441e-08    554.209 -> 8.6306e-09    554.209 -> 8.6306e-09
-findBracket init:    0 -> 1.9441e-08    831.314 -> 5.39479e-09    831.314 -> 5.39479e-09
-findBracket init:    0 -> 1.9441e-08    1246.97 -> 2.40548e-09    1246.97 -> 2.40548e-09
-findBracket init:    0 -> 1.9441e-08    1870.46 -> 5.35411e-10    1870.46 -> 5.35411e-10
-findBracket init:    0 -> 1.9441e-08    2805.69 -> 2.88915e-11    2805.69 -> 2.88915e-11
-findBracket init:    0 -> 1.9441e-08    4208.53 -> 1.22416e-09    4208.53 -> 1.22416e-09
-findBracket init:    0 -> 1.9441e-08    6312.79 -> 4.14884e-08    6312.79 -> 4.14884e-08
-findBracket init:    0 -> 1.9441e-08    9469.19 -> 5.30893e-07    9469.19 -> 5.30893e-07
-findBracket init:    0 -> 1.9441e-08    14203.8 -> 4.53751e-06    14203.8 -> 4.53751e-06
-findBracket init:    0 -> 1.9441e-08    21305.7 -> 3.15269e-05    21305.7 -> 3.15269e-05
-findBracket init:    0 -> 1.9441e-08    31958.5 -> 0.000194681    31958.5 -> 0.000194681
-findBracket (C):    -31958.5 -> 0.00039689    0 -> 1.9441e-08    31958.5 -> 0.000194681
-findBracket:    -31958.5 -> 0.00039689    0 -> 1.9441e-08    31958.5 -> 0.000194681
-parabgoldMinimization:    -31958.5 -> 0.00039689    0 -> 1.9441e-08    31958.5 -> 0.000194681
-parabgoldMinimization:    -31958.5 -> 0.00039689    5462.34 -> 1.39599e-08    31958.5 -> 0.000194681
-Current position and target value:
-x0 = 1.00677    x1 = -1.48958    x2 = 0.49695
-f = 1.39599e-08 +- 1e-05
-
-
-End ConjGrad::findMin() procedure
-
