+++ examples/CMakeLists.txt
@@ -1,0 +1,4 @@
+link_libraries(nfm)
+add_executable(ex1.exe ex1/main.cpp)
+add_executable(ex2.exe ex2/main.cpp)
+add_executable(ex3.exe ex3/main.cpp)
