+++ include/nfm/Noisy1DFunction.hpp
