+++ src/FunProjection1D.cpp
@@ -1,4 +1,4 @@
-#include "FunProjection1D.hpp"
+#include "nfm/FunProjection1D.hpp"
