--- src/nfm
@@ -1,1 +1,0 @@
-./
