+++ test/ut4/main.cpp
@@ -2,11 +2,11 @@
 #include <assert.h>
 #include <math.h>
 
-#include "NoisyFunction.hpp"
-#include "NoisyFunctionValue.hpp"
-#include "1DTools.hpp"
-#include "DynamicDescent.hpp"
-#include "LogNFM.hpp"
+#include "nfm/NoisyFunction.hpp"
+#include "nfm/NoisyFunctionValue.hpp"
+#include "nfm/1DTools.hpp"
+#include "nfm/DynamicDescent.hpp"
+#include "nfm/LogNFM.hpp"
